+++ src/Helper/HelperMaps.h
@@ -1,0 +1,17 @@
+#pragma once
+
+bool IsLoading();
+
+bool IsExplorable();
+
+bool IsOutpost();
+
+bool IsMapReady();
+
+bool IsEndGameEntryOutpost();
+
+bool IsFowEntryOutpost();
+
+bool IsDoa();
+
+bool IsDoaEntryOutpost();
